Refactor connector_mapper.py into ~400-line modules
connector_mapper.py was 1266 lines. Split into a thin orchestrator plus
eight cohesive helpers; the largest new file is 342 lines.

  connector_mapper.py             288  (orchestrator)
  mapper/sm_to_fm_pipeline.py     342  (run() + 7 core _stage_* methods)
  mapper/semantic_match_runner.py 220  (stages 10/11)
  mapper/user_config_matcher.py   154  (stages 7/9 + override normalizer)
  mapper/tco_processor.py         121  (process_tco_information)
  constants/connector_packs.py    116  (premium/commercial frozensets)
  mapper/template_resolver.py      95  (stage 3 SM+FM template fetch/load)
  mapper/v1_to_v2_post_translator  92  (stages 2/13 transformer glue)
  mapper/connector_file_parser.py  69  (parse_connector_file)

External surface is preserved on ConnectorMapper for back-compat:
DISCOVERED_CONFIGS_DIR / SUCCESSFUL_CONFIGS_SUBDIR /
UNSUCCESSFUL_CONFIGS_SUBDIR class constants, parse_connector_file static
method, process_connectors, transform_sm_to_fm, process_tco_information,
connector_pack_type, and the premium_pack_connectors /
commercial_pack_connectors attrs. No callers in cli/, migration/,
output/, or tests had to change.

One incidental fix: parse_connector_file calls os.path.exists() but
'import os' was lost in Phase 3a's dead-import sweep on this branch
(master had it). The new connector_file_parser.py imports os, so the
function actually runs again instead of raising NameError. Tests didn't
exercise that path so this regression wasn't caught.

144/144 tests pass (26 SM->FM integration recordings + 118 unit).
